[#173] test: nanaService 기존 테스트 삭제
